Move canceling tasks on close to the default client
