+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/http/ServiceCall.java
@@ -3,13 +3,6 @@ package com.microsoft.appcenter.http;
 public interface ServiceCall {
 
     /**
-     * Waits if necessary, and then retrieves its status.
-     *
-     * @return true if the call was finished.
-     */
-    boolean ensureFinished();
-
-    /**
      * Cancel the call if possible.
      */
     void cancel();
